+++ include/cone_light_speaker.h
@@ -11,7 +11,7 @@ class ConeLightSpeaker
 {
 private:
   CyberarmSong *m_song = new CyberarmSong();
-  unsigned long m_last_milliseconds = millis();
+  uint32_t m_last_milliseconds = millis();
   bool m_speaker_inited = false;
   bool m_boot_sound_played = false;
   ConeLight *m_cone_light = nullptr;
